delete user data function implemented

+++ server/controller/deleteUserData.js
@@ -1,35 +1,36 @@
 import db from "../db";
 
-export default async function fetchReadme(req, res) {
-    const userId = req.params.id;
+export default async function deleteUserData(userId) {
 	try {
 		const userResponse = await db.query(
-            `SELECT * FROM users WHERE userId = $1`,
-            [userId]
-        );
+			`SELECT * FROM users WHERE userId = $1`,
+			[userId]
+		);
 
-        if(userResponse.rows.length === 0){
-            return res
-            .status(404)
-            .json({ success: false, message: "User not found" });
-        }
+		if (userResponse.rows.length === 0) {
+			return { success: false, message: "User not found" };
+		}
 
 		try {
-			const result = await db.query(
-				`DELETE FROM activities WHERE user_id = $1;
-                DELETE FROM profiles WHERE user_id = $1;
-                DELETE FROM projects WHERE user_id = $1;
-                DELETE FROM pull_requests WHERE user_id = $1;
-                DELETE FROM readmes WHERE user_id = $1;
-                DELETE FROM users WHERE user_id = $1;
-                `,
-				[userId]
-			);
-			return { result: result, message: "successfully added to db" };
+			await db.query("BEGIN");
+			await db.query(`DELETE FROM activities WHERE user_id = $1`, [userId]);
+			await db.query(`DELETE FROM profiles WHERE user_id = $1`, [userId]);
+			await db.query(`DELETE FROM projects WHERE user_id = $1`, [userId]);
+			await db.query(`DELETE FROM pull_requests WHERE user_id = $1`, [userId]);
+			await db.query(`DELETE FROM readmes WHERE user_id = $1`, [userId]);
+			await db.query("COMMIT");
+
+			return {
+				success: true,
+				message: "Related data successfully deleted from the database",
+			};
 		} catch (error) {
-			return { error: error, message: "Cannot connect to db" };
+			await db.query("ROLLBACK");
+			throw new Error(
+				"Failed to delete related data from the database: " + error.message
+			);
 		}
 	} catch (error) {
-		return { error: error, message: "Cannot connect to db" };
+		throw new Error("Database connection error: " + error.message);
 	}
 }
